Make TemplatedView work as IContentView
