+++ src/Controls/src/Core/HandlerImpl/ContentView/ContentView.Impl.cs
@@ -3,6 +3,7 @@
 	public partial class ContentView : IContentView
 	{
 		object IContentView.Content => Content;
+
 		IView IContentView.PresentedContent => ((this as IControlTemplated).TemplateRoot as IView) ?? Content;
 
 		protected override void OnApplyTemplate()
