+++ src/Controls/src/Core/HandlerImpl/RadioButton/RadioButton.Impl.cs
@@ -12,30 +12,6 @@ namespace Microsoft.Maui.Controls
 		object IContentView.Content => ContentAsString();
 #endif
 
-		protected override Size MeasureOverride(double widthConstraint, double heightConstraint)
-		{
-			DesiredSize = this.ComputeDesiredSize(widthConstraint, heightConstraint);
-			return DesiredSize;
-		}
-
-		Size IContentView.CrossPlatformMeasure(double widthConstraint, double heightConstraint)
-		{
-			return this.MeasureContent(widthConstraint, heightConstraint);
-		}
-
-		protected override Size ArrangeOverride(Rect bounds)
-		{
-			Frame = this.ComputeFrame(bounds);
-			Handler?.PlatformArrange(Frame);
-			return Frame.Size;
-		}
-
-		Size IContentView.CrossPlatformArrange(Rect bounds)
-		{
-			this.ArrangeContent(bounds);
-			return bounds.Size;
-		}
-
 		IView IContentView.PresentedContent => ((this as IControlTemplated).TemplateRoot as IView) ?? (Content as IView);
 
 		double IButtonStroke.StrokeThickness => (double)GetValue(BorderWidthProperty);
