+++ src/Controls/src/Core/HandlerImpl/TemplatedView/TemplatedView.Impl.cs
@@ -1,0 +1,38 @@
+﻿#nullable enable
+using Microsoft.Maui.Graphics;
+using Microsoft.Maui.Layouts;
+
+namespace Microsoft.Maui.Controls
+{
+	public partial class TemplatedView : IContentView
+	{
+		object? IContentView.Content => null;
+
+		IView? IContentView.PresentedContent =>
+			(this as IControlTemplated).TemplateRoot as IView;
+
+		protected override Size MeasureOverride(double widthConstraint, double heightConstraint)
+		{
+			DesiredSize = this.ComputeDesiredSize(widthConstraint, heightConstraint);
+			return DesiredSize;
+		}
+
+		Size IContentView.CrossPlatformMeasure(double widthConstraint, double heightConstraint)
+		{
+			return this.MeasureContent(widthConstraint, heightConstraint);
+		}
+
+		protected override Size ArrangeOverride(Rect bounds)
+		{
+			Frame = this.ComputeFrame(bounds);
+			Handler?.PlatformArrange(Frame);
+			return Frame.Size;
+		}
+
+		Size IContentView.CrossPlatformArrange(Rect bounds)
+		{
+			this.ArrangeContent(bounds);
+			return bounds.Size;
+		}
+	}
+}
